COMP: Do not constraint project depending on OpenIGTLink with template-depth of 50
This commit increases the template-depth to 900 to avoid compilation issue
when building application with C++11 against Qt5.

The value of 900 was selected because it is the default with gcc.
See https://gcc.gnu.org/onlinedocs/gcc-4.8.2/gcc/C_002b_002b-Dialect-Options.html#C_002b_002b-Dialect-Options
and https://stackoverflow.com/questions/23374953/why-does-this-exceed-the-maximum-recursive-template-depth#comment35810463_23376156