FIX: make code compile under C++20
In the destructor of the igtlLightObject, a call to
std::uncaught_exception was used. This function was deprecated in C++17
and then removed in C++20, which causes OpenIGTLink to fail compilation
on Visual Studio if C++20 is enabled.
Judging from the comments in the code, the call was unnecessary, so it
was removed.